Update using temp dir in tests
